FileLogger: Add a simple file logger class.
To allow for logging of data to a file, add a simple file logger which
allows for appending of data to file storage.
The class takes a path and the base name for a file. It then prepends a
timestamp to the file. The file to log to can be changed on the fly by
a call to NewFile.
The data is logged by appending to the file and the file is opened and
closed on each case.