Main Activity: Add a logging button and a text view for VID/PID
To allow for the logging of the incoming data from serial port, add a
button which will call the usb services' ToggleLog method and specify a
base name of serial.log, which will result in a file being logged to
in external storage with a <datetime>-serial.log name.
Also add a new text view for writing the PID/VID which is returned from
the usb service when a new usb serial device is connected.